massa grejer

+++ utvecklingsprojekt/articles.php
@@ -15,11 +15,18 @@ if($res->num_rows > 0)
 		{$row->benamning}<br>
 		{$row->utpris_prislista_a}
 		<a href="article_details.php?artikelnr={$row->artikelnr}">Läs mer</a><br>
+END;
+if(isset($_SESSION['userId']))
+{
+	    $content .= <<<END
 		<a href="delete.php?artikelnr={$row->artikelnr}" onclick="return
 		confirm('Är du säker?')">Ta bort artikel</a><br>
 		<a href="edit_article.php?artikelnr={$row->artikelnr}">Redigera artikel</a><br>
-		<hr>
+		
 END;
+}
+        $content .= '<hr>';
+
 	}
 }
