+++ utvecklingsprojekt/delete.php
@@ -1,6 +1,6 @@
 <?php
 include('template.php');
-if(isset($_GET['artikelnr']))
+if(isset($_GET['artikelnr'])&&isset($_SESSION['userId']))
 {
 	$query = <<<END 
 	DELETE FROM artikel
